Extract StatsTable into own file; add title to ReviewedUnreviewedTabs.
