+++ src/uniprotkb/components/statistics/ReviewedUnreviewedTabs.tsx
@@ -1,16 +1,21 @@
 import { Tab, Tabs } from 'franklin-sites';
 
 type Props = {
+  title: string;
   children: [reviewed: React.ReactNode, unreviewed: React.ReactNode];
 };
 
 const ReviewedUnreviewedTabs = ({
+  title,
   children: [reviewed, unreviewed],
 }: Props) => (
-  <Tabs>
-    <Tab title="Reviewed (Swiss-Prot)">{reviewed}</Tab>
-    <Tab title="Unreviewed (TrEMBL)">{unreviewed}</Tab>
-  </Tabs>
+  <>
+    <h3>{title}</h3>
+    <Tabs>
+      <Tab title="Reviewed (Swiss-Prot)">{reviewed}</Tab>
+      <Tab title="Unreviewed (TrEMBL)">{unreviewed}</Tab>
+    </Tabs>
+  </>
 );
 
 export default ReviewedUnreviewedTabs;
